fix(agents): ParseAgent ExtractString numPages/confidence are uint8, not uint256
Grounded against docs.somnia.network/agents/base-agents/llm-parse-website: the live
parse-website agent's selector is ExtractString(...,bool,uint8,uint8). Our uint256 width
produced selector 0x15385dfb instead of the correct 0xc2dd1a7a — every real parse request
would have TimedOut (mock tests didn't catch it; the mock ignores the selector). Fixed in
IParseAgent, SomniaAI.encodeExtractString, ParseAgent.ExtractParams, and the SDK parseAgentAbi.

Also confirmed against official docs (verbatim match, no change needed): IAgentRequester,
IAgentRequesterHandler, Response/Request structs, ResponseStatus enum, deposit formula,
ILlmAgent.inferString/inferNumber. Contracts 131 tests pass; SDK 37 pass.

+++ packages/contracts/src/agents/lib/SomniaAI.sol
@@ -80,6 +80,10 @@ library SomniaAI {
     /// @param resolveUrl  true = domain-search mode (discover pages first); false = direct scrape
     /// @param numPages    pages to read (capped at 1 when resolveUrl == false)
     /// @param confidenceThreshold 0–100 confidence gate below which extraction fails
+    /// @dev numPages/confidenceThreshold are `uint8` to match the live agent's selector
+    ///      `ExtractString(string,string,string[],string,string,bool,uint8,uint8)` — confirmed
+    ///      against docs.somnia.network/agents/base-agents/llm-parse-website. A uint256 here
+    ///      would change the selector and the request would TimeOut against the real agent.
     function encodeExtractString(
         string memory key,
         string memory description,
@@ -87,8 +91,8 @@ library SomniaAI {
         string memory prompt,
         string memory url,
         bool resolveUrl,
-        uint256 numPages,
-        uint256 confidenceThreshold
+        uint8 numPages,
+        uint8 confidenceThreshold
     ) internal pure returns (bytes memory) {
         return abi.encodeWithSelector(
             IParseAgent.ExtractString.selector,
